+++ Assets/Scripts/Models/BusinessModel.cs
@@ -10,7 +10,7 @@ public class BusinessModel
     public int GetBasicPrice => _basicPrice;
     public int GetBaseIncome => _baseIncome; 
     public int GetCurrentLevel => _currentLevel;
-    public float GetCurrentIncome => _currentIncome;
+    public int GetCurrentIncome => _currentIncome;
     public int GetCurrentLevelPrice => _currentLevelPrice;
     public ImprovementModel[] GetImprovemnts => _imrpovements;
     
@@ -18,12 +18,12 @@ public class BusinessModel
     [SerializeField] private int _delayIncome;
     [SerializeField] private int _basicPrice;
     [SerializeField] private int _baseIncome;
-    [SerializeField] private float _currentIncome;
+    [SerializeField] private int _currentIncome;
     [SerializeField] private int _currentLevel;
     [SerializeField] private int _currentLevelPrice;
     [SerializeField] private ImprovementModel[] _imrpovements;
 
-    public void ChangeProfit(float newProfit)
+    public void ChangeProfit(int newProfit)
     {
         _currentIncome = newProfit;
     }
